Implemented more generic enum features
