+++ catowser/catowser/Utils/Features/BaseModels/GenericEnumFeature.swift
@@ -1,0 +1,36 @@
+//
+//  GenericEnumFeature.swift
+//  catowser
+//
+//  Created by Andrey Ermoshin on 26.02.2022.
+//  Copyright © 2022 andreiermoshin. All rights reserved.
+//
+
+import Foundation
+
+protocol EnumDefaultValueSupportable where Self: CaseIterable {
+    var defaultValue: Self { get }
+}
+
+struct GenericEnumFeature<E: CaseIterable & RawRepresentable & EnumDefaultValueSupportable>: EnumFeature where E.RawValue == Int {
+    typealias RawValue = E.RawValue
+    typealias EnumValue = E
+    
+    let wrappedEnumValue: EnumValue
+    
+    var defaultEnumValue: EnumValue {
+        return wrappedEnumValue.defaultValue
+    }
+    
+    var defaultRawValue: RawValue {
+        return defaultEnumValue.rawValue
+    }
+    
+    let key: String
+    
+    init(_ key: String) {
+        // swiftlint:disable:next force_unwrapping
+        wrappedEnumValue = E.allCases.first!
+        self.key = key
+    }
+}
