+++ catowser/catowser/Utils/Features/BaseModels/RawFeatureValue.swift
@@ -1,0 +1,15 @@
+//
+//  RawFeatureValue.swift
+//  catowser
+//
+//  Created by Andrey Ermoshin on 26.02.2022.
+//  Copyright © 2022 andreiermoshin. All rights reserved.
+//
+
+import Foundation
+
+// Raw types of features we support.
+protocol RawFeatureValue {}
+extension Bool: RawFeatureValue {}
+extension Int: RawFeatureValue {}
+extension String: RawFeatureValue {}
